pyproject.toml
[build-system]
requires = [
    "setuptools>=61.0",
    "wheel",
    "setuptools-scm>=8.3.0",
    "semver>=3.0.1"
]
build-backend = "setuptools.build_meta"

[tool.setuptools.dynamic]
version = {attr = "p99_sso_login_proxy.__version__"}

[project]
name = "P99LoginProxy"
dynamic = ["version"]
description = "Proxy for EQ P99 SSO authentication"
readme = "README.md"
requires-python = ">=3.9"
authors = [
    {name = "Adam Harwell", email = "flux.adam@gmail.com"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: GUI",
    "Environment :: Win32 (MS Windows)",
    "Environment :: X11 Applications",
    "Intended Audience :: End Users/Desktop",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.13",
    "Topic :: Games/Entertainment",
    "Topic :: Internet :: Proxy Servers",
]
dependencies = [
    "pycryptodomex>=3.15.0",
    "wxPython>=4.2.0",
    "requests>=2.31.0",
    "pywin32>=306; sys_platform == 'win32'",
    "pystray>=0.19",
    "Pillow>=10.0",
    "semver>=3.0.1",
    "markdown>=3.8",
    "watchdog>=6.0.0",
    "httpx>=0.28.1",
    "websockets>=13.0",
]

[project.scripts]
p99loginproxy = "p99_sso_login_proxy.cmd:main"

[tool.setuptools]
packages = ["p99_sso_login_proxy"]

[tool.setuptools.package-data]
"*" = ["tray_icon*"]

[project.optional-dependencies]
dev = [
    "pyinstaller>=6.13.0",
    "ruff",
    "tox",
]

[tool.pycodestyle]
max-line-length = 120

[tool.basedpyright]
venvPath = "."
venv = ".venv-3.13"

[tool.ruff]
line-length = 120
target-version = "py313"

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "UP",     # pyupgrade
    "B",      # flake8-bugbear
    "SIM",    # flake8-simplify
    "RUF",    # ruff-specific rules
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
